# src/app/swagger-prod.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0",
      "title": "Cutshort",
      "contact": {}
    },
    "host": "joelreuben28.onrender.com/",
    "basePath": "/",
    "securityDefinitions": {
      "Authorization": {
        "type": "apiKey",
        "name": "authorization",
        "in": "header",
        "description": "Enter the token with the `Bearer: ` prefix, e.g. \"Bearer abcde12345\""
      }
    },
    "schemes": [
      "https"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
...
      "/post/{postId}": {
        "patch": {
          "summary": "Update Post",
          "tags": [
            "Post"
          ],
          "operationId": "UpdatePost",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "Body",
              "in": "body",
              "required": true,
              "description": "",
              "schema": {
                "$ref": "#/definitions/UpdatePostRequest"
              }
            },
            {
              "name": "postId",
              "in": "path",
              "required": true,
              "type": "string",
              "description": ""
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "security": [{"Authorization": []}]
        },
        "delete": {
          "summary": "Delete Post",
          "tags": [
            "Post"
          ],
          "operationId": "DeletePost",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "postId",
              "in": "path",
              "required": true,
              "type": "string",
              "description": ""
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "security": [{"Authorization": []}]
        },
        "get": {
          "summary": "Get Post By Id",
          "tags": [
            "Post"
          ],
          "operationId": "GetPostById",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "postId",
              "in": "path",
              "required": true,
              "type": "string",
              "description": ""
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "security": [{"Authorization": []}]
        }
      },
...
      "/user/{userId}": {
        "patch": {
          "summary": "Update User",
          "tags": [
            "User"
          ],
          "operationId": "UpdateUser",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "Body",
              "in": "body",
              "required": true,
              "description": "",
              "schema": {
                "$ref": "#/definitions/UpdateUserRequest"
              }
            },
            {
              "name": "userId",
              "in": "path",
              "required": true,
              "type": "string",
              "description": ""
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "security": [{"Authorization": []}]
        },
        "get": {
          "summary": "Get User",
          "tags": [
            "User"
          ],
          "operationId": "GetUser",
          "deprecated": false,
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "required": true,
              "type": "string",
              "description": ""
            }
          ],
          "responses": {
            "200": {
              "description": "",
              "headers": {}
            }
          },
          "security": [{"Authorization": []}]
        }
      },
...
    },
    "definitions": {
...
      "UpdatePostRequest": {
        "title": "UpdatePostRequest",
        "example": {
          "title": "Title Old",
          "body": "Body Old"
        },
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "body"
        ]
      },
...
      "UpdateUserRequest": {
        "title": "UpdateUserRequest",
        "example": {
          "role": "admin"
        },
        "type": "object",
        "properties": {
          "role": {
            "type": "string"
          }
        },
        "required": [
          "role"
        ]
      }
    },
    "security": [{"Authorization": []}]
  }